Broke up the view model logic into smaller scoped classes. Also fixed navigation issues as well as added disconnect support throughtout app
