+++ dapp/src/main/kotlin/com/walletconnect/dapp/domain/DappDelegate.kt
@@ -3,15 +3,15 @@ package com.walletconnect.dapp.domain
 import android.util.Log
 import com.walletconnect.walletconnectv2.client.WalletConnect
 import com.walletconnect.walletconnectv2.client.WalletConnectClient
+import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
 
 object DappDelegate : WalletConnectClient.DappDelegate {
-    private val _wcEventModels: MutableSharedFlow<WalletConnect.Model> = MutableSharedFlow(extraBufferCapacity = 1)
-    val wcEventModels: SharedFlow<WalletConnect.Model> = _wcEventModels
+    private val TAG: String = DappDelegate::class.java.canonicalName!!
+    private val _wcEventModels: MutableSharedFlow<Pair<Long, WalletConnect.Model?>> = MutableSharedFlow(extraBufferCapacity = 2)
+    val wcEventModels: SharedFlow<Pair<Long, WalletConnect.Model?>> = _wcEventModels
 
-    var selectedPairingTopic: String? = null
-        private set
     var selectedSessionTopic: String? = null
         private set
     var selectedAccountDetails: Triple<String, String, String>? = null
@@ -22,44 +22,47 @@ object DappDelegate : WalletConnectClient.DappDelegate {
     }
 
     override fun onPairingSettled(settledPairing: WalletConnect.Model.SettledPairing) {
-        Log.e(DappDelegate::class.java.canonicalName, "pairing settled")
+        Log.d(TAG, "pairing settled")
     }
 
     override fun onPairingUpdated(pairing: WalletConnect.Model.SettledPairing) {
-        Log.e(DappDelegate::class.java.canonicalName, "pairing updated")
+        Log.d(TAG, "pairing updated")
     }
 
     override fun onSessionApproved(approvedSession: WalletConnect.Model.ApprovedSession) {
-        Log.e(DappDelegate::class.java.canonicalName, "session approved")
+        Log.d(TAG, "session approved")
         selectedSessionTopic = approvedSession.topic
 
-        _wcEventModels.tryEmit(approvedSession)
+        _wcEventModels.tryEmit(System.currentTimeMillis() to approvedSession)
+        _wcEventModels.tryEmit(System.currentTimeMillis() to null)
     }
 
     override fun onSessionRejected(rejectedSession: WalletConnect.Model.RejectedSession) {
-        TODO("Not yet implemented")
+        Log.d(TAG, "session rejected")
+
+        _wcEventModels.tryEmit(System.currentTimeMillis() to rejectedSession)
+        _wcEventModels.tryEmit(System.currentTimeMillis() to null)
     }
 
     override fun onSessionUpdate(updatedSession: WalletConnect.Model.UpdatedSession) {
-        Log.e(DappDelegate::class.java.canonicalName, "session updated")
+        Log.d(TAG, "session updated")
 
-        _wcEventModels.tryEmit(updatedSession)
+        _wcEventModels.tryEmit(System.currentTimeMillis() to updatedSession)
     }
 
     override fun onSessionUpgrade(upgradedSession: WalletConnect.Model.UpgradedSession) {
-        Log.e(DappDelegate::class.java.canonicalName, "session upgrade")
+        Log.d(TAG, "session upgrade")
 
-        _wcEventModels.tryEmit(upgradedSession)
+        _wcEventModels.tryEmit(System.currentTimeMillis() to upgradedSession)
     }
 
     override fun onSessionDelete(deletedSession: WalletConnect.Model.DeletedSession) {
-        Log.e(DappDelegate::class.java.canonicalName, "session deleted")
+        Log.d(TAG, "session deleted")
 
-        _wcEventModels.tryEmit(deletedSession)
-    }
+        selectedAccountDetails = null
+        selectedSessionTopic = null
 
-    fun setSelectedPairingTopicOnConnect(topic: String) {
-        selectedPairingTopic = topic
+        _wcEventModels.tryEmit(System.currentTimeMillis() to deletedSession)
     }
 
     fun setSelectedAccountDetails(accountDetails: String) {
@@ -69,5 +72,6 @@ object DappDelegate : WalletConnectClient.DappDelegate {
 
     fun deselectAccountDetails() {
         selectedAccountDetails = null
+        selectedSessionTopic = null
     }
 }
